deduplicate media toggle functions
